#94 fix/Fix onclick btn dropdown students projects for one jpo with component jpo card & fix little css
