+++ client/src/pages/ressources/jpo/JpoInfo.scss
@@ -17,7 +17,6 @@
     display: flex;
     align-items: center;
     padding: 10px;
-    width: 65%;
 }
 
 .jpo-right-side-section {
@@ -70,7 +69,6 @@
 }
 
 .button-project {
-    width: 35%;
     display: flex;
     padding: 10px;
     justify-content: center;
